refactor: 상태 패턴 적용한 부분 제거
