+++ src/main/java/chess/controller/ChessGameController.java
@@ -63,7 +63,7 @@ public class ChessGameController {
     }
 
     private void move(final ChessGame chessGame, final List<String> commands) {
-        if (!chessGame.isInitialized()) {
+        if (chessGame.isNotInitialized()) {
             throw new IllegalArgumentException("START를 입력해주세요.");
         }
         final MoveDto moveDto = new MoveDto(commands.get(MOVE_SOURCE_INDEX), commands.get(MOVE_TARGET_INDEX));
@@ -72,14 +72,14 @@ public class ChessGameController {
     }
 
     private void status(final ChessGame chessGame) {
-        if (!chessGame.isInitialized()) {
+        if (chessGame.isNotInitialized()) {
             throw new IllegalArgumentException("START를 입력해주세요.");
         }
         OutputView.printStatus(chessGame.getResult());
     }
 
     private void clear(final ChessGame chessGame) {
-        if (!chessGame.isInitialized()) {
+        if (chessGame.isNotInitialized()) {
             throw new IllegalArgumentException("START를 입력해주세요.");
         }
         chessGame.clear();
