+++ src/main/java/chess/service/ChessGame.java
@@ -2,7 +2,6 @@ package chess.service;
 
 import chess.domain.board.Board;
 import chess.domain.board.GameResult;
-import chess.domain.board.InitialState;
 import chess.domain.position.Position;
 import chess.dto.MoveDto;
 import chess.repository.ChessDao;
@@ -13,7 +12,7 @@ public class ChessGame {
     private final ChessDao chessDao;
 
     public ChessGame(final ChessDao chessDao) {
-        this(new InitialState(), chessDao);
+        this(new Board(), chessDao);
     }
 
     public ChessGame(final Board board, final ChessDao chessDao) {
@@ -22,28 +21,32 @@ public class ChessGame {
     }
 
     public void initialize() {
-        board = board.initialize();
+        board.initialize();
         final List<MoveDto> moves = chessDao.findAll();
         for (MoveDto move : moves) {
             final Position source = Position.from(move.getSource());
             final Position target = Position.from(move.getTarget());
-            board = board.move(source, target);
+            board.move(source, target);
         }
     }
 
     public boolean isInitialized() {
-        return board.isInitialized();
+        return !board.isNotInitialized();
+    }
+
+    public boolean isNotInitialized() {
+        return board.isNotInitialized();
     }
 
     public void move(final MoveDto moveDto) {
         final Position sourcePosition = Position.from(moveDto.getSource());
         final Position targetPosition = Position.from(moveDto.getTarget());
-        board = board.move(sourcePosition, targetPosition);
+        board.move(sourcePosition, targetPosition);
         chessDao.save(moveDto);
     }
 
     public void clear() {
-        board = new InitialState();
+        board = new Board();
         chessDao.deleteAll();
     }
