+++ src/test/java/chess/domain/board/GameResultTest.java
@@ -31,8 +31,8 @@ public class GameResultTest {
     @Test
     void 기물의_총_점수를_반환한다() {
         // given
-        Board board = new InitialState();
-        board = board.initialize();
+        Board board = new Board();
+        board.initialize();
         final GameResult result = board.getResult();
 
         // when
@@ -45,11 +45,11 @@ public class GameResultTest {
     @Test
     void 하나의_File에_여러_개의_폰이_존재하는_경우_각폰의_점수가_반이_된다() {
         // given
-        Board board = new InitialState();
-        board = board.initialize();
-        board = board.move(E2, E4);
-        board = board.move(D7, D5);
-        board = board.move(E4, D5);
+        Board board = new Board();
+        board.initialize();
+        board.move(E2, E4);
+        board.move(D7, D5);
+        board.move(E4, D5);
         final GameResult result = board.getResult();
 
         // when
@@ -62,13 +62,13 @@ public class GameResultTest {
     @Test
     void 왕이_잡힌_경우_왕을_잡은쪽이_승리한다() {
         // given
-        Board board = new InitialState();
-        board = board.initialize();
-        board = board.move(E2, E4);
-        board = board.move(E7, E5);
-        board = board.move(D1, H5);
-        board = board.move(F7, F5);
-        board = board.move(H5, E8);
+        Board board = new Board();
+        board.initialize();
+        board.move(E2, E4);
+        board.move(E7, E5);
+        board.move(D1, H5);
+        board.move(F7, F5);
+        board.move(H5, E8);
         final GameResult result = board.getResult();
 
         // when
@@ -81,12 +81,12 @@ public class GameResultTest {
     @Test
     void 왕이_잡히지_않은_경우_점수를_비교하여_결과를_계산한다() {
         // given
-        Board board = new InitialState();
-        board = board.initialize();
-        board = board.move(D2, D4);
-        board = board.move(B8, C6);
-        board = board.move(C1, F4);
-        board = board.move(C6, D4);
+        Board board = new Board();
+        board.initialize();
+        board.move(D2, D4);
+        board.move(B8, C6);
+        board.move(C1, F4);
+        board.move(C6, D4);
         final GameResult result = board.getResult();
 
         // when
@@ -99,8 +99,8 @@ public class GameResultTest {
     @Test
     void 양쪽다_왕이_살아있고_점수가_동일할_경우_EMPTY를_반환한다() {
         // given
-        Board board = new InitialState();
-        board = board.initialize();
+        Board board = new Board();
+        board.initialize();
         final GameResult result = board.getResult();
 
         // when
